Fix nested package data being used as a shared dependency.
When merging dependencies, use the internal name (tpl[0]) rather than the filename
to ensure that nested structures are maintained.

Ensure that extension suffix's are added to dependencies internal name (the first element in a toc).

Update multipackage test to include a test where nested package data is used as a shared dependency.